Add isEqualCaseInsensitive to controller-utils
